requirements.txt
flask==2.0.3
faiss-cpu==1.7.4
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
numpy==1.23.5
spacy==3.5.3